+++ sdk/platform/Order/OrderPlatformValidator.d.ts
@@ -155,6 +155,12 @@ export = OrderPlatformValidator;
  */
 /** @typedef GetRoleBasedActionsParam */
 /**
+ * @typedef GetShipmentBagReasonsParam
+ * @property {string} shipmentId - ID of the bag. An order may contain multiple
+ *   items and may get divided into one or more shipment, each having its own ID.
+ * @property {number} lineNumber - Line number of bag.
+ */
+/**
  * @typedef GetShipmentByIdParam
  * @property {string} [channelShipmentId] - App Shipment Id
  * @property {string} [shipmentId] - Shipment Id
@@ -333,6 +339,8 @@ declare class OrderPlatformValidator {
     static getOrders(): GetOrdersParam;
     /** @returns {GetRoleBasedActionsParam} */
     static getRoleBasedActions(): any;
+    /** @returns {GetShipmentBagReasonsParam} */
+    static getShipmentBagReasons(): GetShipmentBagReasonsParam;
     /** @returns {GetShipmentByIdParam} */
     static getShipmentById(): GetShipmentByIdParam;
     /** @returns {GetShipmentHistoryParam} */
@@ -377,7 +385,7 @@ declare class OrderPlatformValidator {
     static verifyMobileOTP(): VerifyMobileOTPParam;
 }
 declare namespace OrderPlatformValidator {
-    export { AttachOrderUserParam, CheckOrderStatusParam, Click2CallParam, CreateChannelConfigParam, CreateOrderParam, DispatchManifestParam, DownloadBulkActionTemplateParam, DownloadLanesReportParam, EInvoiceRetryParam, FetchCreditBalanceDetailParam, FetchRefundModeConfigParam, GeneratePOSReceiptByOrderIdParam, GetAllowedStateTransitionParam, GetAnnouncementsParam, GetBagByIdParam, GetBagsParam, GetBulkActionTemplateParam, GetBulkShipmentExcelFileParam, GetChannelConfigParam, GetLaneConfigParam, GetOrderByIdParam, GetOrdersParam, GetRoleBasedActionsParam, GetShipmentByIdParam, GetShipmentHistoryParam, GetShipmentReasonsParam, GetShipmentsParam, GetStateTransitionMapParam, GetfiltersParam, InvalidateShipmentCacheParam, OrderUpdateParam, PostShipmentHistoryParam, ProcessManifestParam, ReassignLocationParam, SendSmsNinjaParam, SendUserMobileOTPParam, TrackShipmentParam, UpdateAddressParam, UpdatePackagingDimensionsParam, UpdateShipmentLockParam, UpdateShipmentStatusParam, UpdateShipmentTrackingParam, UploadConsentParam, VerifyMobileOTPParam };
+    export { AttachOrderUserParam, CheckOrderStatusParam, Click2CallParam, CreateChannelConfigParam, CreateOrderParam, DispatchManifestParam, DownloadBulkActionTemplateParam, DownloadLanesReportParam, EInvoiceRetryParam, FetchCreditBalanceDetailParam, FetchRefundModeConfigParam, GeneratePOSReceiptByOrderIdParam, GetAllowedStateTransitionParam, GetAnnouncementsParam, GetBagByIdParam, GetBagsParam, GetBulkActionTemplateParam, GetBulkShipmentExcelFileParam, GetChannelConfigParam, GetLaneConfigParam, GetOrderByIdParam, GetOrdersParam, GetRoleBasedActionsParam, GetShipmentBagReasonsParam, GetShipmentByIdParam, GetShipmentHistoryParam, GetShipmentReasonsParam, GetShipmentsParam, GetStateTransitionMapParam, GetfiltersParam, InvalidateShipmentCacheParam, OrderUpdateParam, PostShipmentHistoryParam, ProcessManifestParam, ReassignLocationParam, SendSmsNinjaParam, SendUserMobileOTPParam, TrackShipmentParam, UpdateAddressParam, UpdatePackagingDimensionsParam, UpdateShipmentLockParam, UpdateShipmentStatusParam, UpdateShipmentTrackingParam, UploadConsentParam, VerifyMobileOTPParam };
 }
 type AttachOrderUserParam = {
     body: OrderPlatformModel.AttachOrderUser;
@@ -654,6 +662,17 @@ type GetOrdersParam = {
     showCrossCompanyData?: boolean;
     customerId?: string;
     orderType?: string;
+};
+type GetShipmentBagReasonsParam = {
+    /**
+     * - ID of the bag. An order may contain multiple
+     * items and may get divided into one or more shipment, each having its own ID.
+     */
+    shipmentId: string;
+    /**
+     * - Line number of bag.
+     */
+    lineNumber: number;
 };
 type GetShipmentByIdParam = {
     /**