+++ sdk/platform/Order/OrderPlatformValidator.js
@@ -181,6 +181,13 @@ const OrderPlatformModel = require("./OrderPlatformModel");
 /** @typedef GetRoleBasedActionsParam */
 
 /**
+ * @typedef GetShipmentBagReasonsParam
+ * @property {string} shipmentId - ID of the bag. An order may contain multiple
+ *   items and may get divided into one or more shipment, each having its own ID.
+ * @property {number} lineNumber - Line number of bag.
+ */
+
+/**
  * @typedef GetShipmentByIdParam
  * @property {string} [channelShipmentId] - App Shipment Id
  * @property {string} [shipmentId] - Shipment Id
@@ -551,6 +558,14 @@ class OrderPlatformValidator {
   /** @returns {GetRoleBasedActionsParam} */
   static getRoleBasedActions() {
     return Joi.object({}).required();
+  }
+
+  /** @returns {GetShipmentBagReasonsParam} */
+  static getShipmentBagReasons() {
+    return Joi.object({
+      shipmentId: Joi.string().allow("").required(),
+      lineNumber: Joi.number().required(),
+    }).required();
   }
 
   /** @returns {GetShipmentByIdParam} */